requirements.txt
python-nest>=4.1.0
rumps>=0.3.0
py2app>=0.19
